requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
executing==2.2.0
idna==3.10
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.53
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.4
python-decouple==3.8
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
stack-data==0.6.3
stripe==11.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
